+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/InvitationController.java
@@ -10,15 +10,18 @@ import lombok.val;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.web.bind.annotation.*;
-import pl.cyfronet.s4e.ex.NotFoundException;
-import pl.cyfronet.s4e.controller.request.CreateInvitationRequest;
+import pl.cyfronet.s4e.controller.request.InvitationRequest;
 import pl.cyfronet.s4e.controller.response.BasicInstitutionResponse;
+import pl.cyfronet.s4e.controller.response.BasicInvitationResponse;
 import pl.cyfronet.s4e.event.OnConfirmInvitationEvent;
+import pl.cyfronet.s4e.event.OnDeleteInvitationEvent;
 import pl.cyfronet.s4e.event.OnRejectInvitationEvent;
 import pl.cyfronet.s4e.event.OnSendInvitationEvent;
+import pl.cyfronet.s4e.ex.NotFoundException;
 import pl.cyfronet.s4e.service.InvitationService;
 
 import javax.validation.Valid;
+import java.util.Set;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static pl.cyfronet.s4e.Constants.API_PREFIX_V1;
@@ -31,6 +34,19 @@ public class InvitationController {
     private final InvitationService invitationService;
     private final ApplicationEventPublisher eventPublisher;
 
+    @Operation(summary = "Fetch all invitations for specific institution")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Successfully retrieved list"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @GetMapping(value = "/institutions/{institution}/invitations")
+    public Set<BasicInvitationResponse> get(@PathVariable("institution") String institutionSlug) throws Exception {
+        return invitationService.findAllBy(institutionSlug, BasicInvitationResponse.class);
+    }
+
     @Operation(summary = "Send invitation email to user")
     @ApiResponses({
             @ApiResponse(responseCode = "200", description = "Invitation was send"),
@@ -40,13 +56,38 @@ public class InvitationController {
             @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
     })
     @PostMapping(value = "/institutions/{institution}/invitations", consumes = APPLICATION_JSON_VALUE)
-    public void send(
-            @RequestBody @Valid CreateInvitationRequest request,
+    public BasicInvitationResponse send(
+            @RequestBody @Valid InvitationRequest request,
             @PathVariable("institution") String institutionSlug
     ) throws Exception {
         val token = invitationService.createInvitationFrom(request.getEmail(), institutionSlug);
         val invitationEvent = new OnSendInvitationEvent(token, LocaleContextHolder.getLocale());
         eventPublisher.publishEvent(invitationEvent);
+
+        return invitationService.findByToken(token, BasicInvitationResponse.class).get();
+    }
+
+    @Operation(summary = "Resend invitation and delete last one")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Invitation was resend"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @PutMapping(value = "/institutions/{institution}/invitation/{token}", consumes = APPLICATION_JSON_VALUE)
+    public BasicInvitationResponse resend(
+            @RequestBody @Valid InvitationRequest request,
+            @PathVariable("institution") String institutionSlug,
+            @PathVariable String token
+    ) throws Exception {
+        val newToken = invitationService.createInvitationFrom(request.getEmail(), institutionSlug);
+        invitationService.deleteBy(token);
+
+        val invitationEvent = new OnSendInvitationEvent(newToken, LocaleContextHolder.getLocale());
+        eventPublisher.publishEvent(invitationEvent);
+
+        return invitationService.findByToken(newTokensudo , BasicInvitationResponse.class).get();
     }
 
     @Operation(summary = "Confirm invitation")
@@ -60,7 +101,7 @@ public class InvitationController {
     public BasicInstitutionResponse confirm(@PathVariable String token) throws NotFoundException {
         invitationService.confirmInvitation(token);
         val institution = invitationService.findInstitutionBy(token, BasicInstitutionResponse.class)
-                .orElseThrow(() -> new NotFoundException("Institution couldn't be found"));;
+                .orElseThrow(() -> new NotFoundException("Institution couldn't be found"));
         val confirmInvitationEvent = new OnConfirmInvitationEvent(token, LocaleContextHolder.getLocale());
         eventPublisher.publishEvent(confirmInvitationEvent);
 
@@ -78,5 +119,19 @@ public class InvitationController {
         invitationService.rejectInvitation(token);
         val rejectInvitationEvent = new OnRejectInvitationEvent(token, LocaleContextHolder.getLocale());
         eventPublisher.publishEvent(rejectInvitationEvent);
+    }
+
+    @Operation(summary = "Delete invitation")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "Invitation have been removed"),
+            @ApiResponse(responseCode = "400", description = "Incorrect request", content = @Content),
+            @ApiResponse(responseCode = "401", description = "Unauthenticated", content = @Content),
+            @ApiResponse(responseCode = "403", description = "Forbidden", content = @Content),
+            @ApiResponse(responseCode = "404", description = "Not found", content = @Content)
+    })
+    @DeleteMapping(value = "/invitations/{token}")
+    public void delete(@PathVariable("token") String token) {
+        val deleteInvitationEvent = new OnDeleteInvitationEvent(token, LocaleContextHolder.getLocale());
+        eventPublisher.publishEvent(deleteInvitationEvent);
     }
 }