+++ s4e-web/src/app/views/settings/people/state/invitation/invitation.store.spec.ts
@@ -1,0 +1,14 @@
+import { InvitationStore } from './invitation.store';
+
+describe('InvitationStore', () => {
+  let store: InvitationStore;
+
+  beforeEach(() => {
+    store = new InvitationStore();
+  });
+
+  it('should create an instance', () => {
+    expect(store).toBeTruthy();
+  });
+
+});