+++ s4e-web/src/app/views/settings/people/state/person/person.query.ts
@@ -8,8 +8,9 @@ import { Person } from './person.model';
 })
 export class PersonQuery extends QueryEntity<PersonState, Person> {
 
-  constructor(protected store: PersonStore) {
-    super(store);
-  }
+  constructor(protected store: PersonStore) { super(store); }
 
+  selectAllWithInvitations$() {
+
+  }
 }