+++ s4e-web/src/app/views/settings/people/state/person/person.service.ts
@@ -2,11 +2,10 @@ import { httpGetRequest$, httpPostRequest$, httpPutRequest$ } from 'src/app/comm
 import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
 import {PersonStore} from './person.store';
-import {S4eConfig} from '../../../../utils/initializer/config.service';
 import {Observable} from 'rxjs';
 import {DEFAULT_GROUP_SLUG, Person } from './person.model';
-import {catchErrorAndHandleStore} from '../../../../common/store.util';
-import {InstitutionQuery} from '../../state/institution/institution.query';
+import { InstitutionQuery } from '../../../state/institution/institution.query';
+import { S4eConfig } from 'src/app/utils/initializer/config.service';
 
 @Injectable({providedIn: 'root'})
 export class PersonService {
@@ -24,10 +23,8 @@ export class PersonService {
 
   deleteMember(userId: string) {
     const institutionSlug = this._institutionQuery.getActiveId();
-    httpPostRequest$(this.http,
-      `${this.config.apiPrefixV1}/institutions/${institutionSlug}/groups/${DEFAULT_GROUP_SLUG}/members/${userId}`, {}, this.store)
-      .subscribe(
-        () => this.store.remove(userId)
-      );
+    const url = `${this.config.apiPrefixV1}/institutions/${institutionSlug}/groups/${DEFAULT_GROUP_SLUG}/members/${userId}`;
+    httpPostRequest$(this.http, url, {}, this.store)
+      .subscribe(() => this.store.remove(userId));
   }
 }