+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The most recent changes are on top, in each type of changes category.
 
 ### Added
 
+- Add operations on created invitations and status of them [#643](https://github.com/cyfronet-fid/sat4envi/issues/643)
+- Create endpoints for existing invitations [#642](https://github.com/cyfronet-fid/sat4envi/issues/642)
+- Confirm or reject invitation with email URLs [#625](https://github.com/cyfronet-fid/sat4envi/issues/625)
+
 ### Changed
 
 ### Fixed